+++ src/main/java/mk/com/theagrodiarybackend/model/exception/PlantationNotFoundException.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class PlantationNotFoundException extends RuntimeException {
 
-    public PlantationNotFoundException(Long plantationIn) {
+    public PlantationNotFoundException(Integer plantationIn) {
         super(String.format("Plantation with id: %d not found", plantationIn));
     }
 }
