+++ src/main/java/mk/com/theagrodiarybackend/repository/ExpenseRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 
     @Query("select e from Expense e where e.expenseId = :expenseId")
-    Optional<Expense> findByExpenseId(Long expenseId);
+    Optional<Expense> findByExpenseId(Integer expenseId);
 }
